# app/Swagger/Schemas/ErrorBadRequest.php
<?php

namespace App\Swagger\Schemas;

use OpenApi\Attributes as OA;

#[OA\Schema(
    schema: 'ErrorBadRequest',
    type: 'object',
    required: ['message'],
    properties: [
        new OA\Property(property: 'message', type: 'string', example: 'Bad request'),
    ],
)]
class ErrorBadRequest {}